Lots of fixes in process of debugging new TKinMet data transfer from phase to class instances.
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/branches/devPhase@2351 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ Modules/m_prfget.cpp
@@ -691,7 +691,7 @@ void TProfil::CalcBcc()
 //             kdTime: current time step (can be changed in TKinMet class)
 // Returns: elapsed calculation time in seconds
 //
-double TProfil::CalcEqstat( double *kdTime, const int kTimeStep, const double kTime )
+double TProfil::CalcEqstat( double *kdTime, const long kTimeStep, const double kTime )
 {
     TSysEq* STat = (TSysEq*)(&aMod[RT_SYSEQ]);
     long int NumIterFIA,  NumIterIPM, NumPrecLoops;
